+++ tests/qss/QssTestDefs.js
@@ -16,10 +16,6 @@
 
 "use strict";
 
-var fluid = require("infusion"),
-    jqUnit = fluid.require("node-jqunit", require, "jqUnit"),
-    gpii = fluid.registerNamespace("gpii");
-
 require("../testUtils.js");
 require("./qssWidget-languageTests.js");
 require("./qssWidget-readAloudTests.js");
@@ -36,124 +32,18 @@ require("./qssService-undoTests.js");
 require("./qssService-saveTests.js");
 require("./qssService-morePanelTests.js");
 require("./qssCommons-tooltipTests.js");
+require("./qssCommons-navigationTests.js");
+require("./qssCommons-translationsTests.js");
 require("./qssCommons-restartWarningTests.js");
 require("./qssCommons-widgetClosingBehaviourTests.js");
 
-// QSS related
-var clickCloseBtn = "jQuery(\".flc-quickSetStrip > div:last-of-type\").click()",
-    clickLanguageBtn = "jQuery(\".flc-quickSetStrip > div:first-of-type\").click()",
-    getQssSettingsList = "(function getItems() { var repeater = fluid.queryIoCSelector(fluid.rootComponent, 'gpii.psp.repeater')[0]; return repeater.model.items; }())";
-
-// QSS Widgets related
-var checkIfMenuWidget = "jQuery('.flc-qssMenuWidget').is(':visible');";
-
-// Generic
-var closeClosableDialog = "jQuery(\".flc-closeBtn\").click()";
+var fluid = require("infusion"),
+    jqUnit = fluid.require("node-jqunit", require, "jqUnit"),
+    gpii = fluid.registerNamespace("gpii");
 
 require("../../src/main/app.js");
 
 fluid.registerNamespace("gpii.tests.qss.testDefs");
-
-gpii.tests.qss.simulateShortcut = function (dialog, shortcut) {
-    dialog.webContents.sendInputEvent({
-        type: shortcut.type || "keyUp",
-        keyCode: shortcut.key,
-        modifiers: shortcut.modifiers || []
-    });
-};
-
-gpii.tests.qss.getFocusedElementIndex = function () {
-    // Note that the elements will be returned in the order in which they appear in the DOM.
-    var qssButtons = jQuery(".fl-qss-button"),
-        focusedElement = jQuery(".fl-focused")[0];
-    return jQuery.inArray(focusedElement, qssButtons);
-};
-
-gpii.tests.qss.pressKey = function (key, modifiers) {
-    return {
-        funcName: "gpii.tests.qss.simulateShortcut",
-        args: [
-            "{that}.app.qssWrapper.qss.dialog",
-            {
-                key: key,
-                modifiers: modifiers
-            }
-        ]
-    };
-};
-
-gpii.tests.qss.assertFocusedElementIndex = function (expectedIndex) {
-    return {
-        task: "gpii.test.invokeFunctionInWebContentsDelayed",
-        args: [
-            "{that}.app.qssWrapper.qss.dialog",
-            gpii.tests.qss.getFocusedElementIndex,
-            100
-        ],
-        resolve: "jqUnit.assertEquals",
-        resolveArgs: [
-            "The correct button in the QSS is focused",
-            expectedIndex,
-            "{arguments}.0"
-        ]
-    };
-};
-
-gpii.tests.qss.clearFocusedElement = function () {
-    jQuery(".fl-qss-button").removeClass("fl-focused fl-highlighted");
-};
-
-var qssSettingsCount = 16;
-
-var navigationSequence = [
-    {
-        func: "{that}.app.tray.events.onTrayIconClicked.fire"
-    },
-    // No focused element at first
-    gpii.tests.qss.assertFocusedElementIndex(-1),
-    // When the right arrow is pressed, the first button in the QSS will be focused.
-    gpii.tests.qss.pressKey("Right"),
-    gpii.tests.qss.assertFocusedElementIndex(0),
-    gpii.tests.qss.pressKey("Tab"),
-    gpii.tests.qss.assertFocusedElementIndex(1),
-    gpii.tests.qss.pressKey("Left"),
-    gpii.tests.qss.assertFocusedElementIndex(0),
-    gpii.tests.qss.pressKey("Tab", ["Shift"]),
-    gpii.tests.qss.assertFocusedElementIndex(qssSettingsCount - 1),
-    gpii.tests.qss.pressKey("Tab", ["Shift"]),
-    gpii.tests.qss.assertFocusedElementIndex(qssSettingsCount - 2),
-    // This part of the test is commented because the button is disabled temporary (GPII-3773) and cannot be focused.
-    // Some tests may be removed or parts of them re-used in the future.
-    // gpii.tests.qss.pressKey("Up"),
-    // gpii.tests.qss.assertFocusedElementIndex(qssSettingsCount - 3),
-    // gpii.tests.qss.pressKey("Down"),
-    // gpii.tests.qss.assertFocusedElementIndex(qssSettingsCount - 2),
-    // gpii.tests.qss.pressKey("Left"),
-    // gpii.tests.qss.assertFocusedElementIndex(qssSettingsCount - 4),
-    // gpii.tests.qss.pressKey("Up"),
-    // gpii.tests.qss.assertFocusedElementIndex(qssSettingsCount - 4),
-    // This part of the test is commented because the button is disabled temporary (GPII-3773) and cannot be focused.
-    // Some tests may be removed or parts of them re-used in the future.
-    // gpii.tests.qss.pressKey("Right"),
-    // gpii.tests.qss.assertFocusedElementIndex(qssSettingsCount - 3),
-    // Manually clear the focused state in order to test the Arrow Left behavior when
-    // there is no focused element.
-    {
-        task: "gpii.test.invokeFunctionInWebContentsDelayed",
-        args: [
-            "{that}.app.qssWrapper.qss.dialog",
-            gpii.tests.qss.clearFocusedElement,
-            100
-        ],
-        resolve: "fluid.identity"
-    }, { // Close the QSS and the PSP
-        func: "gpii.test.executeJavaScriptInWebContents",
-        args: [
-            "{that}.app.qssWrapper.qss.dialog",
-            clickCloseBtn
-        ]
-    }
-];
 
     // TODO this could be used instead (of the previous)
     //{ // ! should send info to broker
@@ -235,79 +125,6 @@ var qssSettingMessagesFixture = {
         }
     }
 };
-
-/*
- * A mock of the settings message
- */
-var expectedTranslatedSettings = {
-    "http://registry\\.gpii\\.net/common/language": {
-        schema: {
-            "title": "Избери Език",
-            "enum": [
-                "Анг",
-                "Арм",
-                "Кит",
-                "Китс",
-                "Кор",
-                "Рус",
-                "Исп"
-            ]
-        },
-        tooltip: "Some different tooltip"
-    },
-    "save": {
-        schema: {
-            "title": "Запиши ме"
-        }
-    },
-    "close": {
-        schema: {
-            "title": "Затвори ме"
-        }
-    },
-    "appTextZoom": {
-        schema: {
-            title: "Приближи този текст"
-        },
-        tooltip: "Some tooltip",
-        tip: "Some helpful tip",
-        widget: {
-            footerTip: "Hopefully something helpful as the tip"
-        }
-    }
-};
-
-
-var crossQssTranslations = [
-    /*
-     * Tests for the QSS i18n behaviour
-     */
-    { // trigger locale change
-        funcName: "{that}.app.applier.change",
-        args: ["locale", "bg"]
-    },
-    { // ... and check whether the settings' text has been updated in the main
-        funcName: "gpii.test.assertLeftHandDeep",
-        args: [
-            "QSS should have its settings translated with locale change",
-            expectedTranslatedSettings,
-            "@expand:gpii.test.objectArrayToHash({that}.app.qssWrapper.model.settings, path)"
-        ]
-    }, { // ... and in the renderer as well
-        task: "gpii.test.executeJavaScriptInWebContents",
-        args: [
-            "{that}.app.qssWrapper.qss.dialog",
-            getQssSettingsList
-        ],
-        resolve: "gpii.test.assertLeftHandDeep",
-        resolveArgs: [
-            "QSS dialog should have its settings translated with locale change",
-            expectedTranslatedSettings,
-            "@expand:gpii.test.objectArrayToHash({arguments}.0, path)"
-        ]
-    }
-];
-
 
 var unorderedInstalledLangsFixture = {
     raw: {
@@ -426,7 +243,7 @@ var qssInstalledLanguages = [
 
 gpii.tests.qss.testDefs = {
     name: "QSS Widget integration tests",
-    expect: 66,
+    expect: 74,
     config: {
         configName: "gpii.tests.dev.config",
         configPath: "tests/configs"
@@ -466,10 +283,8 @@ gpii.tests.qss.testDefs = {
             listener: "jqUnit.assert",
             args: ["QSS has initialized successfully"]
         }],
-        // For no particular reason the tests work properly in this sequence
-        // navigationSequence,
-        // crossQssTranslations,
         // Commons
+        gpii.tests.qss.navigationTests,
         qssInstalledLanguages,
         gpii.tests.qss.tooltipTests,
         gpii.tests.qss.widgetClosingBehaviourTests,
@@ -489,6 +304,7 @@ gpii.tests.qss.testDefs = {
         gpii.tests.qss.screenCaptureTests,
         gpii.tests.qss.officeSimplifyTests,
         gpii.tests.qss.menuTests,
-        gpii.tests.qss.stepperindicatorsSequence
+        gpii.tests.qss.stepperindicatorsSequence,
+        gpii.tests.qss.translationsTests
     )
 };
